requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
pydantic==2.5.0
pydantic-settings==2.1.0
annotated-types==0.7.0
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.1.1

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
